requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pymupdf==1.23.7
transformers==4.35.2
faiss-cpu==1.7.4
openai==1.3.0
python-multipart==0.0.6
torch==2.1.1
numpy==1.26.2
rouge-score
pandas==2.1.4
openpyxl==3.1.2
markdown==3.5.1 